Impact and vulnerability can be between 0 and 1. Legends were removed from the graph.

+++ src/main/java/com/nikitinho/riskmanager/domain/Risk.java
@@ -70,10 +70,10 @@ public class Risk {
     private Project project;
 
     @JsonView(Views.FullRisk.class)
-    private Integer probability;
+    private Double probability;
 
     @JsonView(Views.FullRisk.class)
-    private Integer impact;
+    private Double impact;
 
     @JsonView(Views.FullRisk.class)
     private Integer riskRate;
